LayoutTests/fast/viewport/viewport-49-expected.txt

CONSOLE MESSAGE: line 3: Viewport width or height set to physical device width, try using "device-width" constant instead for future compatibility.
CONSOLE MESSAGE: line 3: Viewport height or height set to physical device height, try using "device-height" constant instead for future compatibility.
viewport size 320x352 scale 1.000000 with limits [1.000000, 5.000000]
